Refactor insert_router_import method into controller generator
